+++ core/src/agents/workflows/github/index.ts
@@ -1,0 +1,3 @@
+export * from './githubAgent.js';
+export * from './types.js';
+export * from './prompts.js';
