+++ core/src/agents/workflows/slack/index.ts
@@ -1,0 +1,3 @@
+export * from './slackAgent.js';
+export * from './types.js';
+export * from './prompts.js';
