+++ core/src/agents/workflows/twitter/index.ts
@@ -1,0 +1,3 @@
+export * from './twitterAgent.js';
+export * from './types.js';
+export * from './prompts.js';
