+++ core/src/services/vectorDb/index.ts
@@ -1,0 +1,2 @@
+export * from './vectorDBPool.js';
+export * from './VectorDB.js';
